refactor(reservation-app): 좌석 가용성 조회의 Seat 포트 참조 변경
- SeatAvailabilityService가 Room 패키지로 이동한 SeatReader를 참조하도록 변경
- 가용 좌석 테스트가 Room 연관관계를 가진 Seat fixture를 사용하도록 변경
- 점유 locator를 fixture에서 생성된 Seat locator 기준으로 검증하도록 정리

- 테스트:
  - AvailableSeats 필터링 테스트 갱신
  - FindAvailableSeatsUseCase와 FindSeatStatusesUseCase 테스트 갱신
