+++ reservation/reservation-application/src/test/java/com/seatliberator/seatliberator/reservation/availability/application/service/FindAvailableSeatsUseCaseTest.java
@@ -5,8 +5,9 @@ import com.seatliberator.seatliberator.reservation.availability.application.port
 import com.seatliberator.seatliberator.reservation.availability.application.port.in.result.AvailableSeatResult;
 import com.seatliberator.seatliberator.reservation.book.application.contract.OccupancySeatLocatorFinder;
 import com.seatliberator.seatliberator.reservation.book.application.contract.OccupancySeatRangeFinder;
+import com.seatliberator.seatliberator.reservation.domain.fixture.RoomFixture;
 import com.seatliberator.seatliberator.reservation.domain.fixture.SeatFixture;
-import com.seatliberator.seatliberator.reservation.seat.application.port.out.SeatReader;
+import com.seatliberator.seatliberator.reservation.room.application.port.out.SeatReader;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -17,7 +18,6 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import java.time.Instant;
 import java.util.List;
 
-import static com.seatliberator.seatliberator.reservation.domain.fixture.SeatLocatorFixture.createLocator;
 import static com.seatliberator.seatliberator.reservation.domain.fixture.TestSupport.fixedClock;
 import static com.seatliberator.seatliberator.reservation.domain.fixture.TimeRangeFixture.createRange;
 import static org.assertj.core.api.Assertions.assertThat;
@@ -47,29 +47,28 @@ public class FindAvailableSeatsUseCaseTest {
     @Test
     @DisplayName("특정 시간 및 방에서 예약할 수 있는 좌석을 반환한다")
     void return_available_seats() {
+        // given
         var roomId = "room-1";
+        var room = new RoomFixture.Builder().roomId(roomId).build();
+
+        var seatBuilder = new SeatFixture.Builder().room(room).createdAt(now);
+        var seatA = seatBuilder.copy().seatId("A").build();
+        var seatB = seatBuilder.copy().seatId("B").build();
+        var seatC = seatBuilder.copy().seatId("C").build();
+        var seats = List.of(seatA, seatB, seatC);
+
         var range = createRange();
 
-        var seatBuilder = new SeatFixture.Builder().createdAt(now);
-
-        var locatorA = createLocator(roomId, "A");
-        var locatorB = createLocator(roomId, "B");
-        var locatorC = createLocator(roomId, "C");
-
-        var locators = List.of(locatorA, locatorB, locatorC);
-
-        var seats = locators.stream()
-                .map(locator -> seatBuilder.copy().locator(locator).build())
-                .toList();
-
+        // when
         when(seatReader.findByRoomId(roomId))
                 .thenReturn(seats);
         when(occupancySeatLocatorFinder.find(roomId, range))
-                .thenReturn(List.of(locatorA));
+                .thenReturn(List.of(seatA.getLocator()));
 
         var query = new FindAvailableSeatQuery(roomId, range);
         var result = useCase.find(query);
 
+        // then
         assertThat(result)
                 .extracting(AvailableSeatResult::seatId)
                 .containsExactlyInAnyOrder("B", "C");
@@ -81,15 +80,19 @@ public class FindAvailableSeatsUseCaseTest {
     @Test
     @DisplayName("방에 좌석이 없으면 빈 리스트를 반환하고 예약은 조회하지 않는다")
     void return_empty_seats_and_skip_reservation_query_when_no_seats_exists_in_room() {
+        // given
         var roomId = "room-1";
+        var room = new RoomFixture.Builder().roomId(roomId).build();
         var range = createRange();
 
+        // when
         when(seatReader.findByRoomId(roomId))
                 .thenReturn(List.of());
 
         var query = new FindAvailableSeatQuery(roomId, range);
         var result = useCase.find(query);
 
+        // then
         assertThat(result).isEmpty();
 
         verify(seatReader).findByRoomId(roomId);
